Privacy Policy
Privacy Policy of Mobile Android app Remove Photo Background:-

We (VsInfo Tech) respect the privacy of everyone using this app (Remove Photo Background). Please check the following regarding the use of information.

Photo Data:-
The processing of the photos is done only inside your device. Remove Photo Background app does not send your photos to the outside of your device.
Our apps do not collect face data and do not send them to the outside of your device. Our apps never share them with third parties.

Advertisement on App:-
 We may use Google AdMob advertisement in this app. For the information to be acquired, purpose of use, provision to third parties, please check the following page.
https://policies.google.com/privacy
http://www.google.com/intl/en/policies/technologies/ads/


Ad distributors may use cookies and ad IDs to display ads.
It is possible to limit acquisition of information by the following method.
<Android>
  Launch "Settings" app -> Check "Google" -> "Ads" -> "Opt out of Ads Personalization"

Temporary Files:-
Temporary files of photos may be saved in the storage area of the application. These files will not be sent to the outside the apps. These files will be deleted from the device when the app is deleted.

Support Pages:-
  The linked "Support Pages" are created on Google's service. Please check the following page for the information, purpose of use, provision to third parties etc. that Google's service obtains.
https://policies.google.com/technologies/partner-sites
https://policies.google.com/privacy

EMail:-
When sending e-mail to us, e-mail address, title, text, etc. are acquired. We will use it only for problem solving and reply.

App Permissions:-
Internet
- for displaying advertisements.
- for loading and saving photos.